requirements.txt
flask
azure-storage-blob
azure-search-documents
openai
python-dotenv